-Fixing UI Layout - Logam Mulia Pick Data State - Pin Screen Add Fix & Add Notification At Wrong PIN - Added Custom Camera Features

+++ lib/models/detail_pengajuan.dart
@@ -83,7 +83,7 @@ class Pengajuan {
   String? caption;
   String? type;
   int? totalAngsuran;
-  double? angsuranPerBulan;
+  int? angsuranPerBulan;
   String? tenor;
   String? status;
   String? keterangan;
